Always write the token to the .docker/config.json file
A dirty monkey-patchy fix.
Maybe it will not work for multiple registries or simultaneous builds on the same machine, but works when you have only one registry.
It seems like it's easy to fix it further in order to support simultaneous jobs, but I just don't have this scenario yet.